move default text annotation in proper location and other fixes
