+++ web/client/utils/__tests__/DrawSupportUtils-test.js
@@ -7,7 +7,7 @@
  */
 
 const expect = require('expect');
-const {boundsToOLExtent, fromLeafletFeatureToQueryform} = require('../DrawSupportUtils');
+const {boundsToOLExtent, fromLeafletFeatureToQueryform, isCompletePolygon} = require('../DrawSupportUtils');
 const L = require('leaflet');
 
 describe('LocaleUtils', () => {
@@ -44,4 +44,15 @@ describe('LocaleUtils', () => {
         expect(convertedBounds.length).toBe(4);
     });
 
+    it('test isCompletePolygon defaults', () => {
+        const polygonCoords1 = [[[1, 1], [2, 2]]];
+        const polygonCoords2 = [[[1, 1], [2, 2], [1, 1]]];
+        const polygonCoords3 = [[[1, 1], [2, 2], [3, 3], [1, 1]]];
+        const polygonCoords4 = [[[1, 1], [2, undefined], [3, 3], [1, 1]]];
+        expect(isCompletePolygon()).toBe(false);
+        expect(isCompletePolygon(polygonCoords1)).toBe(false);
+        expect(isCompletePolygon(polygonCoords2)).toBe(false);
+        expect(isCompletePolygon(polygonCoords3)).toBe(true);
+        expect(isCompletePolygon(polygonCoords4)).toBe(false);
+    });
 });
